requirements.txt
Django==1.8.5
django-allauth==0.23.0
django-localflavor==1.1
funcsigs==0.4
geopy==1.11.0
jsonfield==1.0.3
oauthlib==1.0.3
pbr==1.8.1
Pillow==2.3.0
psycopg2==2.6.1
python-openid==2.2.5
requests==2.8.1
requests-oauthlib==0.5.0
selenium==2.48.0
six==1.10.0
wheel==0.24.0
